+++ esm6/index.js
@@ -1,9 +1,12 @@
 export * from './client';
 export * from './frame-impl';
+export * from './i-frame';
+export * from './i-message';
 export * from './parser';
 export * from './stomp-config';
 export * from './stomp-headers';
 export * from './stomp-subscription';
+export * from './i-transaction';
 export * from './types';
 export * from './versions';
 // Compatibility code