+++ esm6/parser.js
@@ -149,7 +149,10 @@ export class Parser {
             return;
         }
         if (byte === LF) {
-            this._results.headers.push([this._headerKey, this._consumeTokenAsUTF8()]);
+            this._results.headers.push([
+                this._headerKey,
+                this._consumeTokenAsUTF8(),
+            ]);
             this._headerKey = undefined;
             this._onByte = this._collectHeaders;
             return;
@@ -185,7 +188,12 @@ export class Parser {
     }
     _retrievedBody() {
         this._results.binaryBody = this._consumeTokenAsRaw();
-        this.onFrame(this._results);
+        try {
+            this.onFrame(this._results);
+        }
+        catch (e) {
+            console.log(`Ignoring an exception thrown by a frame handler. Original exception: `, e);
+        }
         this._initState();
     }
     // Rec Descent Parser helpers